- update to today snapshot (20091110)
Changed files:
    cliapps.patch -> 1.2
    dokuwiki-config.patch -> 1.8
    dokuwiki-geshi.patch -> 1.4
    dokuwiki-http_auth-option.patch -> 1.4
    dokuwiki-mail-headerencodequotes.patch -> 1.3
    dokuwiki-nice_exit.patch -> 1.2
    dokuwiki-paths.patch -> 1.5
    dokuwiki.spec -> 1.57

+++ dokuwiki.spec
@@ -1,14 +1,12 @@
-%define		subver	2009-02-14b
-%define		ver	%(echo %{subver} | tr -d -)
 Summary:	PHP-based Wiki webapplication
 Summary(pl.UTF-8):	Aplikacja WWW Wiki oparta na PHP
 Name:		dokuwiki
-Version:	%{ver}
-Release:	7
+Version:	20091110
+Release:	0.5
 License:	GPL v2
 Group:		Applications/WWW
-Source0:	http://www.splitbrain.org/_media/projects/dokuwiki/%{name}-%{subver}.tgz
-# Source0-md5:	c75c4781b8698041c3c9b6b0fec2ac2e
+Source0:	http://dev.splitbrain.org/download/snapshots/dokuwiki-latest.tgz
+# Source0-md5:	9da3e5c3a7b40b5a229b38e52b6fcf70
 Source1:	%{name}-apache.conf
 Source2:	%{name}-lighttpd.conf
 Source3:	jude.png
@@ -17,8 +15,8 @@ Source4:	eventum.gif
 # Source4-md5:	cac3d0f82591a33eda2afa8ae5fe65cb
 Source5:	http://forum.skype.com/style_emoticons/skype/skype.png
 # Source5-md5:	25c355be038267dc9fdb724b628000b9
+Patch66:	%{name}-config.patch
 Patch0:		%{name}-paths.patch
-Patch1:		%{name}-config.patch
 Patch2:		%{name}-mysqlauth.patch
 Patch3:		%{name}-config-allow-require.patch
 Patch4:		%{name}-geshi.patch
@@ -33,8 +31,6 @@ Patch12:	%{name}-mailthreads.patch
 Patch13:	%{name}-media-directlink.patch
 Patch14:	interwiki-outputonly.patch
 Patch15:	simplepie.patch
-Patch16:	cliapps.patch
-Patch17:	plugin_exists-check.patch
 Patch18:	install.patch
 URL:		http://wiki.splitbrain.org/wiki:dokuwiki
 BuildRequires:	rpmbuild(macros) >= 1.520
@@ -92,9 +88,8 @@ po pierwszej instalacji. Potem należy go odinstalować, jako że
 pozostawienie plików instalacyjnych mogłoby być niebezpieczne.
 
 %prep
-%setup -q -n %{name}-2009-02-14
+%setup -q -n %{name}
 %patch0 -p1
-%patch1 -p1
 %patch2 -p1
 %patch3 -p1
 %patch4 -p1
@@ -109,9 +104,9 @@ pozostawienie plików instalacyjnych mogłoby być niebezpieczne.
 %patch13 -p1
 %patch14 -p1
 %patch15 -p1
-%patch16 -p1
-%patch17 -p1
 %patch18 -p1
+
+%patch66 -p1
 
 find -name _dummy | xargs rm
 rm -f lib/index.html lib/plugins/index.html
@@ -130,8 +125,18 @@ rm -f inc/SimplePie.php
 # our plugins dir is not writable anyway, nothing to convert
 rm -rf lib/plugins/upgradeplugindirectory
 
+# flash source
+rm -rf lib/_fla
+
 # cleanup backups after patching
 find . '(' -name '*~' -o -name '*.orig' ')' -print0 | xargs -0 -r -l512 rm -f
+
+%build
+md5=$(md5sum -b conf/dokuwiki.php | awk '{print $1}')
+if ! grep $md5 install.php; then
+	: update %{name}-config.patch oudated
+	exit 1
+fi
 
 %install
 rm -rf $RPM_BUILD_ROOT
@@ -143,9 +148,9 @@ cp -a inc/* $RPM_BUILD_ROOT%{_appdir}/inc
 cp -a lib/* $RPM_BUILD_ROOT%{_appdir}/lib
 cp -a conf/* $RPM_BUILD_ROOT%{_sysconfdir}
 cp -a data/* $RPM_BUILD_ROOT%{_localstatedir}
-install %{SOURCE1} $RPM_BUILD_ROOT%{_sysconfdir}/apache.conf
-install %{SOURCE1} $RPM_BUILD_ROOT%{_sysconfdir}/httpd.conf
-install %{SOURCE2} $RPM_BUILD_ROOT%{_sysconfdir}/lighttpd.conf
+cp -a %{SOURCE1} $RPM_BUILD_ROOT%{_sysconfdir}/apache.conf
+cp -a %{SOURCE1} $RPM_BUILD_ROOT%{_sysconfdir}/httpd.conf
+cp -a %{SOURCE2} $RPM_BUILD_ROOT%{_sysconfdir}/lighttpd.conf
 touch $RPM_BUILD_ROOT%{_sysconfdir}/acronyms.local.conf
 touch $RPM_BUILD_ROOT%{_sysconfdir}/entities.local.conf
 touch $RPM_BUILD_ROOT%{_sysconfdir}/interwiki.local.conf
@@ -157,7 +162,7 @@ touch $RPM_BUILD_ROOT%{_sysconfdir}/smileys.local.conf
 touch $RPM_BUILD_ROOT%{_sysconfdir}/userstyle.css
 
 ln $RPM_BUILD_ROOT%{_appdir}/lib/images/interwiki/{dokubug,bug}.gif
-cp -a %{SOURCE3} $RPM_BUILD_ROOT%{_appdir}/lib/images/fileicons
+cp -a %{SOURCE3} $RPM_BUILD_ROOT%{_appdir}/lib/images/fileicons/jide.png
 cp -a %{SOURCE4} $RPM_BUILD_ROOT%{_appdir}/lib/images/interwiki/eventum.gif
 cp -a %{SOURCE5} $RPM_BUILD_ROOT%{_appdir}/lib/images/interwiki/skype.gif
 
@@ -278,9 +283,11 @@ exit 0
 %{_appdir}/lib/plugins/acl/pix
 %dir %{_appdir}/lib/plugins/config
 %{_appdir}/lib/plugins/config/*.*
+%{_appdir}/lib/plugins/config/images
 %{_appdir}/lib/plugins/config/settings
 %dir %{_appdir}/lib/plugins/plugin
 %{_appdir}/lib/plugins/plugin/*.*
+%{_appdir}/lib/plugins/plugin/classes
 %dir %{_appdir}/lib/plugins/revert
 %{_appdir}/lib/plugins/revert/*.*
 %dir %{_appdir}/lib/plugins/usermanager
@@ -306,12 +313,10 @@ exit 0
 %dir %attr(770,root,http) %{_localstatedir}/meta
 %dir %attr(770,root,http) %{_localstatedir}/pages
 %dir %attr(770,root,http) %{_localstatedir}/pages/wiki
-%dir %attr(770,root,http) %{_localstatedir}/pages/playground
 %dir %attr(770,root,http) %{_localstatedir}/tmp
 %attr(660,root,http) %config(noreplace,missingok) %verify(not md5 mtime size) %{_localstatedir}/media/wiki/dokuwiki-128.png
 %attr(660,root,http) %config(noreplace,missingok) %verify(not md5 mtime size) %{_localstatedir}/pages/wiki/dokuwiki.txt
 %attr(660,root,http) %config(noreplace,missingok) %verify(not md5 mtime size) %{_localstatedir}/pages/wiki/syntax.txt
-%attr(660,root,http) %config(noreplace,missingok) %verify(not md5 mtime size) %{_localstatedir}/pages/playground/playground.txt
 
 %files setup
 %defattr(644,root,root,755)
